각 직업별 vote 기능 구현 완료

+++ app/src/main/java/com/codefair/lawfeedback/model/UpdateArticleTO.java
@@ -3,8 +3,8 @@ package com.codefair.lawfeedback.model;
 public class UpdateArticleTO {
 
     private Long jobId;
-    private Long good;
-    private Long bad;
+    private Integer good;
+    private Integer bad;
 
     public Long getJobId() {
         return jobId;
@@ -14,23 +14,23 @@ public class UpdateArticleTO {
         this.jobId = jobId;
     }
 
-    public Long getGood() {
+    public Integer getGood() {
         return good;
     }
 
-    public void setGood(Long good) {
+    public void setGood(Integer good) {
         this.good = good;
     }
 
-    public Long getBad() {
+    public Integer getBad() {
         return bad;
     }
 
-    public void setBad(Long bad) {
+    public void setBad(Integer bad) {
         this.bad = bad;
     }
 
-    public UpdateArticleTO(Long jobId, Long good, Long bad) {
+    public UpdateArticleTO(Long jobId, Integer good, Integer bad) {
         this.jobId = jobId;
         this.good = good;
         this.bad = bad;
